--- app/src/main/java/ru/dudar/findfilms/domain/Themoviesgenres/Genres.kt
@@ -1,11 +1,0 @@
-package ru.dudar.findfilms.domain.Themoviesgenres
-
-import com.google.gson.annotations.SerializedName
-
-
-data class Genres (
-
-    @SerializedName("id") var id : Int,
-    @SerializedName("name") var name : String
-
-)
