+++ app/src/main/java/ru/dudar/findfilms/domain/Themoviesgenres/TheMovieGenres.kt
@@ -9,3 +9,10 @@ data class TheMovieGenres (
     @SerializedName("genres") var genres : ArrayList<Genres>
 
 )
+
+data class Genres (
+
+    @SerializedName("id") var id : Int,
+    @SerializedName("name") var name : String
+
+)
